// ClientApp/app/models/perfil-cliente.ts
export class PerfilCliente {
    celularDdd: string;
    celularNumero: string;
    cep: string;
    cidade: string;
    bairro: string;
    classeSetor: string;
    complemento: string;
    cpf: string;
    dtNascimento: Date;
    estadoCivil: string;
    idade: number;
    idSexo: string;
    idEstado: string;
    idEstadoCivil: number;
    idFaixaPatrimonio: number;
    idNacionalidade: number;
    logradouro: string;
    nome: string;
    nomeMae: string;
    nomePai: string;
    numero: string;
    sexo: string;
    telefoneDdd: string;
    telefoneNumero: string;
    telefoneRamal: string;
    nacionalidade: string;
    idUfExpedidor: string;

    celularDddAlterado: boolean;
    celularNumeroAlterado: boolean;
    cepAlterado: boolean;
    cidadeAlterado: boolean;
    classeSetorAlterado: boolean;
    complementoAlterado: boolean;
    cpfAlterado: boolean;
    dtNascimentoAlterado: boolean;
    estadoCivilAlterado: boolean;
    idadeAlterado: boolean;
    idSexoAlterado: boolean;
    idEstadoAlterado: boolean;
    idEstadoCivilAlterado: boolean;
    idFaixaPatrimonioAlterado: boolean;
    idNacionalidadeAlterado: boolean;
    logradouroAlterado: boolean;
    nomeAlterado: boolean;
    nomeMaeAlterado: boolean;
    numeroAlterado: boolean;
    sexoAlterado: boolean;
    telefoneDddAlterado: boolean;
    telefoneNumeroAlterado: boolean;
    telefoneRamalAlterado: boolean;
    nacionalidadeAlterado: boolean;
    idUfExpedidorAlterado: boolean;
}